+++ CHANGELOG.md
@@ -5,7 +5,20 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [4.0.0] 2025-07
+## [4.1.0] 2025-07-21
+
+### Added
+
+- New module: "SQL"
+    - `DatabaseManager`, A user-friendly context manager for handling SQLite database operations.
+
+### Changed
+
+- PSO_optimization
+    - Updated logic to save best results to a .csv directly and incrementally, instead of loading them all in memory.
+    - Supports saving data to a SQL database, working with `SQL.DatabaseManager` under the hood.
+
+## [4.0.0] 2025-07-19
 
 ### Added